requirements.txt
numpy>=1.21.0
scipy>=1.7.0
cma>=3.0.0
gymnasium>=0.29.0
ale-py>=0.8.0
PyYAML>=6.0
ray>=2.0.0
matplotlib>=3.5.0
scienceplots>=2.0.0
tqdm>=4.62.0
tabulate>=0.9.0
